Move utils and event under Apache
