+++ disco-java-agent-web/src/main/java/software/amazon/disco/agent/web/apache/event/ApacheEventFactory.java
@@ -13,9 +13,9 @@
  *   permissions and limitations under the License.
  */
 
-package software.amazon.disco.agent.web.apache.httpclient.event;
+package software.amazon.disco.agent.web.apache.event;
 
-import software.amazon.disco.agent.web.apache.httpclient.utils.HttpRequestAccessor;
+import software.amazon.disco.agent.web.apache.utils.HttpRequestAccessor;
 
 /**
  * Create our private events, so that listeners do not have public access to them
