+++ disco-java-agent-web/src/main/java/software/amazon/disco/agent/web/apache/utils/HttpResponseAccessor.java
@@ -13,7 +13,7 @@
  *   permissions and limitations under the License.
  */
 
-package software.amazon.disco.agent.web.apache.httpclient.utils;
+package software.amazon.disco.agent.web.apache.utils;
 
 import software.amazon.disco.agent.web.MethodHandleWrapper;
