+++ disco-java-agent-web/src/main/java/software/amazon/disco/agent/web/apache/utils/MethodInterceptionCounter.java
@@ -13,7 +13,7 @@
  *   permissions and limitations under the License.
  */
 
-package software.amazon.disco.agent.web.apache.httpclient.utils;
+package software.amazon.disco.agent.web.apache.utils;
 
 import java.util.concurrent.atomic.AtomicInteger;
